Minor refactor of directories
