+++ app/app.py
@@ -3,50 +3,19 @@ import datetime
 import json
 import logging
 import os
-import re
 
-import gh_md_to_html
 import pytz
 from flasgger import Swagger
 from flask import Flask, request
 
 from app.pesu import PESUAcademy
 from pydantic import ValidationError
-from app.models import RequestModel, ResponseModel
+from app.models import ResponseModel
+import app.util as util
 
 IST = pytz.timezone("Asia/Kolkata")
 app = Flask(__name__)
 pesu_academy = PESUAcademy()
-
-
-def convert_readme_to_html():
-    """
-    Convert the README.md file to HTML and save it as README.html so that it can be rendered on the home page.
-    """
-    logging.info("Beginning conversion of README.md to HTML...")
-    readme_content = open("README.md").read().strip()
-    readme_content = re.sub(r":\w+: ", "", readme_content)
-    with open("README_tmp.md", "w") as f:
-        f.write(readme_content)
-    html = gh_md_to_html.main("README_tmp.md").strip()
-    with open("README.html", "w") as f:
-        f.write(html)
-    logging.info("README.md converted to HTML successfully.")
-
-
-def validate_input(data: dict) -> RequestModel:
-    """
-    Validate the input provided by the user
-    :param data: The input data provided by the user
-    :return: The validated data as a RequestModel object
-    """
-    logging.info(
-        f"Validating input: {data.get('username')=}, password={'*****' if data.get('password') else None}, "
-        f"profile={data.get('profile')}, fields={data.get('fields')}"
-    )
-    validated_data = RequestModel.model_validate(data)
-    logging.info("Input validation successful. All parameters are valid.")
-    return validated_data
 
 
 @app.route("/readme")
@@ -67,7 +36,7 @@ def readme():
     try:
         if "README.html" not in os.listdir():
             logging.info("README.html does not exist. Beginning conversion...")
-            convert_readme_to_html()
+            util.convert_readme_to_html()
         logging.info("Rendering README.html...")
         with open("README.html") as f:
             output = f.read()
@@ -217,7 +186,7 @@ def authenticate():
     # Validate the input provided by the user
     try:
         logging.info("Received authentication request. Beginning input validation...")
-        validated_data = validate_input(request.json)
+        validated_data = util.validate_input(request.json)
         username = validated_data.username
         password = validated_data.password
         profile = validated_data.profile
