+++ tests/unit/test_readme.py
@@ -9,8 +9,7 @@ def test_convert_readme_to_html_writes_html():
     with patch("builtins.open", m_open):
         with patch("gh_md_to_html.main") as mock_gh_md_to_html:
             mock_gh_md_to_html.return_value = "<h1>Title</h1>"
-            app_module.convert_readme_to_html()
-    # Now you can assert open was called to write "README_tmp.md" and "README.html"
+            app_module.util.convert_readme_to_html()
     m_open.assert_any_call("README_tmp.md", "w")
     m_open.assert_any_call("README.html", "w")
     handle = m_open()
@@ -22,8 +21,7 @@ def test_readme_raises_exception(monkeypatch):
         raise Exception("fail")
 
     monkeypatch.setattr(app_module.os, "listdir", lambda: [])
-    monkeypatch.setattr(app_module, "convert_readme_to_html", raise_exception)
-
+    monkeypatch.setattr(app_module.util, "convert_readme_to_html", raise_exception)
     response, status_code = app_module.readme()
     assert status_code == 500
     assert "Error occurred" in response
